fixed wish list adaptive for mobile, preloader moved to utils, refactored terms of service page - moved code to components
