+++ client/src/components/TermsOfService/ListOfTerms.js
