+++ client/src/components/TermsOfService/styles.js
