+++ client/src/pages/Categories/index.js
@@ -5,5 +5,4 @@ export { default as CategoriesRadio } from "./CategoriesRadio.jsx";
 export { default as TopFilter } from "./Catalog Product List/TopFilter.jsx";
 export { default as ShowQuantity } from "./Catalog Product List/ShowQuantity.jsx";
 export { default as SortBy } from "./Catalog Product List/SortBy.jsx";
-export { default as Preloader } from "./Catalog Product List/Preloader.jsx";
 export { default as CatalogProductList } from "./Catalog Product List/CatalogProductList.jsx";
